Renamed GuiApp.PrismAvalonia to GuiNet5. Upgraded GuiNet6 NuGet packages to the latest Prism.Avalonia v8.1.97
