+++ sandbox/GuiNet5/Assets/avalonia-logo.ico
