+++ sandbox/GuiNet5/ViewModels/DashboardViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
 using Prism.Commands;
 
-namespace GuiApp.PrismAvalonia.ViewModels
+namespace GuiNet5.ViewModels
 {
   public class DashboardViewModel : ViewModelBase
   {
