+++ sandbox/GuiNet5/ViewModels/ShellWindowViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace GuiApp.PrismAvalonia.ViewModels
+﻿namespace GuiNet5.ViewModels
 {
   public class ShellWindowViewModel : ViewModelBase
   {
