+++ sandbox/GuiNet5/Views/DashboardView.axaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="GuiApp.PrismAvalonia.Views.DashboardView">
+             x:Class="GuiNet5.Views.DashboardView">
   <DockPanel>
     <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">
       <TextBlock Text="Welcome to Prism.Avalonia!" />
