+++ sandbox/GuiNet5/Views/DashboardView.axaml.cs
@@ -1,17 +1,14 @@
-using Avalonia;
+﻿using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace GuiApp.PrismAvalonia.Views
+namespace GuiNet5.Views
 {
-  public partial class ShellWindow : Window
+  public partial class DashboardView : UserControl
   {
-    public ShellWindow()
+    public DashboardView()
     {
       InitializeComponent();
-#if DEBUG
-            this.AttachDevTools();
-#endif
     }
 
     private void InitializeComponent()
