+++ sandbox/GuiNet5/Views/SettingsView.axaml
@@ -5,7 +5,7 @@
              xmlns:prism="http://prismlibrary.com/"
              prism:ViewModelLocator.AutoWireViewModel="True"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-             x:Class="GuiApp.PrismAvalonia.Views.SettingsView">
+             x:Class="GuiNet5.Views.SettingsView">
   <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">>
     <Label>Welcome to Avalonia!</Label>
     <TextBlock Text="{Binding Title}" />
