+++ sandbox/GuiNet5/Views/ShellWindow.axaml
@@ -2,12 +2,12 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:core="clr-namespace:GuiApp.PrismAvalonia;assembly=GuiApp.PrismAvalonia"
-        xmlns:local="using:GuiApp.PrismAvalonia.Views"
+        xmlns:core="clr-namespace:GuiNet5;assembly=GuiNet5"
+        xmlns:local="using:GuiNet5.Views"
         xmlns:prism="http://prismlibrary.com/"
         prism:ViewModelLocator.AutoWireViewModel="True"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="GuiApp.PrismAvalonia.Views.ShellWindow"
+        x:Class="GuiNet5.Views.ShellWindow"
         Background="Transparent"
         Icon="/Assets/avalonia-logo.ico"
         Width="900" Height="700"
